requirements.txt
fastapi
uvicorn
pika
python-dotenv
pydantic
websockets
redis
flake8
numpy
pandas
scikit-learn
disutils
python-multipart
torch
torchvision
torchaudio